+++ CHANGELOG.md
@@ -25,7 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [APPS] Removed the example AEs from the CSE. This makes the CSE a bit smaller and also removes a big dependency to the non-portable psutils package. They will be available in  separate project in the future.
 
 ### Fixed
-
+- [CSE] When the CSE is reset then the statistics are now reset as well.
 
 
 ## [0.8.0] - 2021-07-27